refactor: separate CTR->GCM migration tests to a separate test file
